BIG-16439 Refactor the events
