+++ src/events/currency-selector.js
@@ -1,10 +1,9 @@
-import BcEvents from './index';
-import _ from 'lodash';
+import BaseEvents from './base';
 
-export default class CurrencySelectorEvents extends BcEvents {
+export default class CurrencySelectorEvents extends BaseEvents {
 
     /**
-     *
+     * @Constructor
      * @param {object} options
      */
     constructor(options) {
